pyproject.toml
[project]
name = "dzencode-test-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bleach>=6.3.0",
    "channels-redis>=4.3.0",
    "channels[daphne]>=4.3.1",
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=23.0.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
